Implement code for solver functions
Implemented code for major and minor diagonal functions. The functions
by checking index of top row and going out of row length to find
diagonals that do not start somewhere in the top row. It will
check with a native function where the space checked exists
before using it to avoid strange outputs